convenience usage of ScannerConfiguration
IP address is the only mandatory argument, so make it a param of the constructor to
always have a valid config. The builder is implicitly convertible into a configuration,
to save the user from typing build() in every situation.

This simplifies the main.cpp into more readable and convenient:
```
ScannerV2 scanner(ScannerConfigurationBuilder(SCANNER_IP).scanRange(ScanRange{ ANGLE_START, ANGLE_END }),
                  laserScanCallback);
```